+++ src/cdk-experimental/scrolling/auto-size-virtual-scroll.ts
@@ -8,10 +8,13 @@
 
 import {coerceNumberProperty} from '@angular/cdk/coercion';
 import {ListRange} from '@angular/cdk/collections';
+import {
+  CdkVirtualScrollViewport,
+  VIRTUAL_SCROLL_STRATEGY,
+  VirtualScrollStrategy
+} from '@angular/cdk/scrolling';
 import {Directive, forwardRef, Input, OnChanges} from '@angular/core';
 import {Observable} from 'rxjs';
-import {VIRTUAL_SCROLL_STRATEGY, VirtualScrollStrategy} from './virtual-scroll-strategy';
-import {CdkVirtualScrollViewport} from './virtual-scroll-viewport';
 
 
 /**
