+++ src/cdk-experimental/scrolling/public-api.ts
@@ -7,8 +7,4 @@
  */
 
 export * from './auto-size-virtual-scroll';
-export * from './fixed-size-virtual-scroll';
 export * from './scrolling-module';
-export * from './virtual-for-of';
-export * from './virtual-scroll-strategy';
-export * from './virtual-scroll-viewport';
