+++ src/cdk-experimental/scrolling/scrolling-module.ts
@@ -8,23 +8,10 @@
 
 import {NgModule} from '@angular/core';
 import {CdkAutoSizeVirtualScroll} from './auto-size-virtual-scroll';
-import {CdkFixedSizeVirtualScroll} from './fixed-size-virtual-scroll';
-import {CdkVirtualForOf} from './virtual-for-of';
-import {CdkVirtualScrollViewport} from './virtual-scroll-viewport';
 
 
 @NgModule({
-  exports: [
-    CdkAutoSizeVirtualScroll,
-    CdkFixedSizeVirtualScroll,
-    CdkVirtualForOf,
-    CdkVirtualScrollViewport,
-  ],
-  declarations: [
-    CdkAutoSizeVirtualScroll,
-    CdkFixedSizeVirtualScroll,
-    CdkVirtualForOf,
-    CdkVirtualScrollViewport,
-  ],
+  exports: [CdkAutoSizeVirtualScroll],
+  declarations: [CdkAutoSizeVirtualScroll],
 })
 export class ScrollingModule {}
