+++ src/cdk-experimental/scrolling/tsconfig-build.json
@@ -1,8 +1,7 @@
 {
   "extends": "../tsconfig-build",
   "files": [
-    "public-api.ts",
-    "./typings.d.ts"
+    "public-api.ts"
   ],
   "angularCompilerOptions": {
     "annotateForClosureCompiler": true,
