+++ src/cdk/scrolling/BUILD.bazel
@@ -7,11 +7,19 @@ ng_module(
   name = "scrolling",
   srcs = glob(["**/*.ts"], exclude=["**/*.spec.ts"]),
   module_name = "@angular/cdk/scrolling",
+  assets = [":virtual-scroll-viewport.css"] + glob(["**/*.html"]),
   deps = [
+    "//src/cdk/coercion",
+    "//src/cdk/collections",
     "//src/cdk/platform",
     "@rxjs",
   ],
   tsconfig = "//src/cdk:tsconfig-build.json",
+)
+
+sass_binary(
+  name = "virtual_scroll_viewport_scss",
+  src = "virtual-scroll-viewport.scss",
 )
 
 ts_library(
