+++ src/cdk/scrolling/fixed-size-virtual-scroll.ts
