+++ src/cdk/scrolling/scrolling-module.ts
@@ -8,11 +8,24 @@
 
 import {PlatformModule} from '@angular/cdk/platform';
 import {NgModule} from '@angular/core';
+import {CdkFixedSizeVirtualScroll} from './fixed-size-virtual-scroll';
 import {CdkScrollable} from './scrollable';
+import {CdkVirtualForOf} from './virtual-for-of';
+import {CdkVirtualScrollViewport} from './virtual-scroll-viewport';
 
 @NgModule({
   imports: [PlatformModule],
-  exports: [CdkScrollable],
-  declarations: [CdkScrollable],
+  exports: [
+    CdkFixedSizeVirtualScroll,
+    CdkScrollable,
+    CdkVirtualForOf,
+    CdkVirtualScrollViewport,
+  ],
+  declarations: [
+    CdkFixedSizeVirtualScroll,
+    CdkScrollable,
+    CdkVirtualForOf,
+    CdkVirtualScrollViewport,
+  ],
 })
-export class ScrollDispatchModule {}
+export class ScrollingModule {}
