+++ src/cdk/scrolling/tsconfig-build.json
@@ -1,7 +1,8 @@
 {
   "extends": "../tsconfig-build",
   "files": [
-    "public-api.ts"
+    "public-api.ts",
+    "./typings.d.ts"
   ],
   "angularCompilerOptions": {
     "annotateForClosureCompiler": true,
