+++ src/cdk/scrolling/viewport-ruler.spec.ts
@@ -1,5 +1,5 @@
 import {TestBed, inject, fakeAsync, tick} from '@angular/core/testing';
-import {ScrollDispatchModule} from './public-api';
+import {ScrollingModule} from './public-api';
 import {ViewportRuler} from './viewport-ruler';
 import {dispatchFakeEvent} from '@angular/cdk/testing';
 
@@ -24,7 +24,7 @@ describe('ViewportRuler', () => {
   veryLargeElement.style.height = '6000px';
 
   beforeEach(() => TestBed.configureTestingModule({
-    imports: [ScrollDispatchModule],
+    imports: [ScrollingModule],
     providers: [ViewportRuler]
   }));
