--- src/cdk/scrolling/viewport.md
@@ -1,3 +1,0 @@
-### Viewport
-
-A utility for getting the bounds of the browser viewport.
