+++ src/cdk/scrolling/virtual-scroll-strategy.ts
