+++ src/cdk/scrolling/virtual-scroll-viewport.scss
