+++ src/demo-app/demo-material-module.ts
@@ -6,18 +6,19 @@
  * found in the LICENSE file at https://angular.io/license
  */
 
-import {ScrollingModule} from '@angular/cdk-experimental/scrolling';
 import {DialogModule} from '@angular/cdk-experimental/dialog';
 import {DragDropModule} from '@angular/cdk-experimental/drag-drop';
+import {ScrollingModule as ExperimentalScrollingModule} from '@angular/cdk-experimental/scrolling';
 import {A11yModule} from '@angular/cdk/a11y';
 import {CdkAccordionModule} from '@angular/cdk/accordion';
 import {BidiModule} from '@angular/cdk/bidi';
-import {TextFieldModule} from '@angular/cdk/text-field';
 import {ObserversModule} from '@angular/cdk/observers';
 import {OverlayModule} from '@angular/cdk/overlay';
 import {PlatformModule} from '@angular/cdk/platform';
 import {PortalModule} from '@angular/cdk/portal';
+import {ScrollingModule} from '@angular/cdk/scrolling';
 import {CdkTableModule} from '@angular/cdk/table';
+import {TextFieldModule} from '@angular/cdk/text-field';
 import {CdkTreeModule} from '@angular/cdk/tree';
 import {NgModule} from '@angular/core';
 import {
@@ -112,6 +113,7 @@ import {
     PlatformModule,
     PortalModule,
     ScrollingModule,
+    ExperimentalScrollingModule,
     DialogModule,
     DragDropModule,
   ]
